feat/ebs-gp3: add basic support to gp3 volumes
Add basic gp3 validation on the provider to support changing the default
control-plane volumes on AWS IPI.

The basic support means that this change will only make the terraform
validate gp3 volumes. The full support should allow adding parameters
like volume IOPS and throughput, which is completely new in General Purpose
volumes (introduced in gp3) - and not covered here as it needs more changes
on this openshift's fork.

References:
- https://github.com/openshift/installer/pull/5239
- https://github.com/openshift/installer/pull/5328